Update styles and content in index.html
In hero-text, put the letter "Data" the same span and format as "Code, and Creativity". Don't emphasize "Data"
Reduce font size of "Selected Works" in hero-text, and move it closer to the spectrum-indicator
Change the map default scale to 1 (transform: translate3d(299.023px, 230.994px, 0px) scale(1);)
On vertical device, when project-card is clicked, expand it to the same ratio as how it expand on a laptop.