Update project card styles and email protection
In mobile view, the project-info should hide when not clicked. Only shows when clicked. Then, get rid of project-title, only keep project-number and project-category, with project-number shown on top-left, project-category shows on bottom-left. This statement is only for mobile view. Nothing will change in laptop view.
In mobile view, when clicked on project-card, expand the card vertically